Screen
Added screen class to keep track of which pixels are set and which
aren't.  A chip8 screen is 64 pixels wide and 32 pixels tall.  Due to
the way java 2d arrays work, we use the y-axis first, and the x-axis
second.  This is why x and y are switched in the setPixel and isPixelSet
methods.  Added test class as well.