Added main class
Added Chip8 class which will act at the main class which binds
everything together.  Added static integers to represent screen size and
replaced any instances of the integers in the code with the variables.
Made Screen extend JPanel, so it can be drawn on.  Added repaint call to
the drawing opcode so the screen will refresh each time the opcode is
called.